Relocate tensor type utilities.
